+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [1.4.1] - 2018-02-21
+## [1.5.0] - 2018-02-21
+
+### Added
+- A possibility to filter Flexible Content layouts by field name.
 
 ### Fixed
 - A few minor bug fixes on various things.